Commit 4
Fixed the following error:

Explanation

Understanding the problem
The error message indicates that you're trying to load a JavaScript module (using <script type="module">) but the server is sending it with the wrong MIME type: text/jsx instead of a JavaScript MIME type like text/javascript or application/javascript. Browsers enforce strict MIME type checking for modules as part of the HTML specification, to prevent security vulnerabilities. In this case, the browser is refusing to load the module, since it is not considered valid JavaScript. The browser is seeing the file as a JSX file, but that is not valid JavaScript.